+++ src/app.py
@@ -1,14 +1,46 @@
+import os
 from fastapi import FastAPI
 from fastapi.responses import JSONResponse
 from app_data import __app_name__, __version__, __date__
 import asyncio
+from core.dispatcher_apprise import DispatcherApprise
 
 from api.common.response_model.successful_request import SuccessfulRequest
+from api.schemas.apprise_test_provider import AppriseTestService
+
 from watchdog import Watchdog
-from config.config import Config
+from config.config import Config, get_configmap
 
 app = FastAPI()
-configHelper = Config()
+
+configHelper = None
+
+
+def load_user_config():
+    global configHelper
+
+    if str(os.getenv('ENV')).lower() == 'development':
+
+        cm = get_configmap(namespace=os.getenv('K8S_VELERO_UI_NAMESPACE', 'velero-ui'),
+                           configmap_name='velero-watchdog-config')
+
+        # Update environment variables
+        for key, value in cm.items():
+            if key != 'K8S_INCLUSTER_MODE' and key != 'PROCESS_KUBE_CONFIG' and key != 'PROCESS_LOAD_KUBE_CONFIG':
+                os.environ[key] = value
+
+    cm = get_configmap(namespace=os.getenv('K8S_VELERO_UI_NAMESPACE', 'velero-ui'),
+                       configmap_name='velero-watchdog-user-config')
+    if cm:
+
+        # Update environment variables
+        for key, value in cm.items():
+            os.environ[key] = value
+
+    configHelper = Config()
+
+
+load_user_config()
 
 app.watchdog_daemon = Watchdog(daemon=True)
 app.task = asyncio.create_task(app.watchdog_daemon.run())
@@ -31,20 +63,22 @@ async def info():
     return JSONResponse(content=response.toJSON(), status_code=200)
 
 
-@app.get("/restart",
-         tags=['System'],
-         summary='Restart service')
+@app.post("/restart",
+          tags=['System'],
+          summary='Restart service')
 async def restart():
     app.task.cancel()
+    load_user_config()
+    app.watchdog_daemon = Watchdog(daemon=True)
     app.task = asyncio.create_task(app.watchdog_daemon.run())
     res = {'restarted': 'Done!'}
     response = SuccessfulRequest(payload=res)
     return JSONResponse(content=response.toJSON(), status_code=200)
 
 
-@app.get("/send-report",
-         tags=['Run'],
-         summary='Send report')
+@app.post("/send-report",
+          tags=['Run'],
+          summary='Send report')
 async def report():
     wd = Watchdog(daemon=False)
     await wd.run()
@@ -53,7 +87,7 @@ async def report():
     return JSONResponse(content=response.toJSON(), status_code=200)
 
 
-@app.get("/get-config",
+@app.get("/environment",
          tags=['System'],
          summary='Get the current configuration')
 async def get_config():
@@ -62,17 +96,12 @@ async def get_config():
     return JSONResponse(content=response.toJSON(), status_code=200)
 
 
-@app.get("/send-test-notification",
-         tags=['Run'],
-         summary='Send a test message to verify channel settings')
-async def send_test_notification(email: bool = True,
-                                 telegram: bool = True,
-                                 slack: bool = True):
-    wd = Watchdog(daemon=False)
-    await wd.run(test_notification=True,
-                 test_email=email,
-                 test_telegram=telegram,
-                 test_slack=slack)
-    res = {'sent': 'Done!'}
+@app.post("/test-service",
+          tags=['Run'],
+          summary='Send a test message to verify channel settings')
+async def send_test_notification(provider: AppriseTestService):
+    dispatcher_test = DispatcherApprise(queue=None, dispatcher_config=None, test_configs=provider.config)
+    success = await dispatcher_test.send_msgs(message='Test message success sent!', test_message=True)
+    res = {'success': success}
     response = SuccessfulRequest(payload=res)
     return JSONResponse(content=response.toJSON(), status_code=200)